+++ packages/origin.js/src/contract-service.js
@@ -1,7 +1,7 @@
-import ListingsRegistryContract from "../build/contracts/ListingsRegistry.json"
-import ListingContract from "../build/contracts/Listing.json"
-import PurchaseContract from "../build/contracts/Purchase.json"
-import UserRegistryContract from "../build/contracts/UserRegistry.json"
+import ListingsRegistryContract from "../../contracts/build/contracts/ListingsRegistry.json"
+import ListingContract from "../../contracts/build/contracts/Listing.json"
+import PurchaseContract from "../../contracts/build/contracts/Purchase.json"
+import UserRegistryContract from "../../contracts/build/contracts/UserRegistry.json"
 import bs58 from "bs58"
 import contract from "truffle-contract"
 import promisify from "util.promisify"
